Add files via upload
Today, everything is virtualized, and there is a large
volume of textual data that is difficult to examine for various
purposes. There are two types of text summarization: extractive
and abstractive. In extractive, the model extracts the main points,
whereas in abstractive, the model generates the summaries. Text
summarization is a one-time approach for comprehending and
evaluating vast amounts of data included in text and files. For
its packages and methods, it is best implemented using NLP.
For automatic text summarization, reinforcement learning is also
proposed, with neural networks used to estimate the Q value.
Rouge is used to assess the model’s performance. In this case,
NLTK (natural language toolkit) is one of the most powerful
libraries in NLP, and spacy is used to analyse and summarise a
significant amount of data.
Index Terms—Extractive, Abstractive, Rouge, Reinforcement
Learning , NLTK, Spacy.